Change FixId struct to use usize for indexing Vecs
<!-- ps-id: fc7d7c26-02ff-4172-9743-e9708f35453e -->

+++ src/commands/fix.rs
@@ -2,8 +2,8 @@ use crate::config::Config;
 
 #[derive(Debug)]
 pub struct FixId {
-    pub project_id: u8,
-    pub fixme_id: u8,
+    pub project_id: usize,
+    pub fixme_id: usize,
 }
 
 pub fn fix(_conf: &mut Config, id: FixId) {
